% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeHePlot.R
\name{deepTimeHePlot}
\alias{deepTimeHePlot}
\title{deepTimeHePlot}
\usage{
deepTimeHePlot(deepTimeProdDF, ribbon = FALSE)
}
\arguments{
\item{deepTimeProdDF}{Structured dataframe that is the result of the deepTimeProd() function}

\item{ribbon}{Optional geom_ribbon() around the min/max of the Monte Carlo model}
}
\description{
Produces a ggplot2 of the cumulative amount of helium produced by samples over the specified geologic time in the deepTimeProd() model.
}
\details{
Takes in the resulting dataframe from deepTimeProd() function and plots the cumulative amount of helium generated by the sample over the duration of the model. Can plot an optional geom_ribbon() around the min/max of the Monte Carlo model.
}
\examples{
data("monteDataLithCat")
df <- monteProd(lithCat,1000)
dfTime <- deepTimeProd(df,1,100,1)
deepTimeH2Plot(dfTime)

}
