% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeProd.R
\name{deepTimeProd}
\alias{deepTimeProd}
\title{deepTimeProd}
\usage{
deepTimeProd(monteProdDF, startAge, endAge, stepAge)
}
\arguments{
\item{monteProdDF}{Output dataframe of a Monte Carlo model}

\item{startAge}{Starting age (youngest age) of the model in millions of years (i.e., a value of 50 is fifty million years)}

\item{endAge}{Ending age (oldest age) of the model in millions of years (i.e., a value of 100 is one hundred million years)}

\item{stepAge}{Step of the model}
}
\description{
Creates a cumulative He/H2 generation model backwards through geologic time, incorporating radioactive decay of U, Th, and K. Production rates (H2total, Hetotal) are in rates of mols gas / km3 rock / year following Warr et al. (2023).
}
\details{
This model is programmed in units of millions of years
}
\examples{
data("monteDataLithCat")
df <- monteProd(lithCat,1000)
deepTimeProd(df,1,100,1)

}
