% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepTimeSum.R
\name{deepTimeSum}
\alias{deepTimeSum}
\title{deepTimeSum}
\usage{
deepTimeSum(deepTimeProdDF)
}
\arguments{
\item{deepTimeProdDF}{Output dataframe of the deepTimeProd()}
}
\description{
Summarizes the output of deepTimeProd() function into cumulative min/max/mean of produced hydrogen and helium over the specified model duration
}
\examples{
data("monteDataLithCat")
df <- monteProd(lithCat,1000)
dfTime <- deepTimeProd(df,1,100,1)
deepTimeSum(dfTime)

}
