% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteH2Plot.R
\name{monteH2Plot}
\alias{monteH2Plot}
\title{monteH2Plot}
\usage{
monteH2Plot(modelDF, ribbon = FALSE)
}
\arguments{
\item{modelDF}{Output dataframe of Monte Carlo model}

\item{ribbon}{Boolean parameter that controls a geom_ribbon() of the min/max production rate of the samples}
}
\description{
Creates a ggplot2 that scales the hydrogen production rate of samples with increasing source rock volume (1km3 to 1000km3)
}
\examples{
data("monteDataLithCat")
df <- monteProd(lithCat,1000)
monteH2Plot(df, ribbon=TRUE)

List references
}
