% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteProd.R
\name{monteProd}
\alias{monteProd}
\title{monteProd}
\usage{
monteProd(sampDF, numGen)
}
\arguments{
\item{sampDF}{Structured sample dataframe}

\item{numGen}{Number of Monte Carlo simulations}
}
\description{
Monte Carlo model of Warr et al (2023) implemented in R.
}
\details{
The monteProd() function takes a structured dataframe and grabs data from the dataframe by column name. As such, look at example data (lithCat and rockProps) for structure. This outputs a dataframe including all Monte Carlo trials for all input samples. Production rates (H2total, Hetotal) are in rates of mols gas / m3 rock / year following Warr et al. (2023).
}
\examples{
data("monteDataLithCat")
monteProd(lithCat,1000)

}
