% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteSum.R
\name{monteSum}
\alias{monteSum}
\title{monteSum}
\usage{
monteSum(modelDF)
}
\arguments{
\item{modelDF}{Monte Carlo model dataframe produced by the monteProd() function.}
}
\description{
Returns a summarized dataframe (one row per sample) from the output dataframe of the monteProd() function.
}
\details{
This function uses a "groupby() %>% summaize()" to produce a min/max/mean/standard deviation for both the hydrogen production and helium production of the input samples.  Production rates (columns ending in H2 and He) are in rates of mols gas / m3 / year following Warr et al. (2023).
}
\examples{
data("monteDataLithCat")
df <- monteProd(lithCat,1000)
monteSum(df)

}
