% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-rockProps.R
\docType{data}
\name{rockProps}
\alias{rockProps}
\title{Example of structured dataframe for radiolysis Monte Carlo simulation with rock property data}
\format{
A data frame formated for
}
\usage{
data(monteDataRockProps)
}
\description{
This is an example dataframe structured for the monteProd() function of Radiolysis that uses a rock property distributions.

\itemize{
  \item Sample. Sample name
  \item rockDenMin. Minimum value of rock density (g/m3) distribution used in simulation
  \item rockDenMinMax. Maximum value of rock density (g/m3) distribution used in simulation
  \item rockDenMinMean. Mean value of rock density (g/m3) distribution used in simulation
  \item rockDenMinSD. Standard deviation of rock density (g/m3) distribution used in simulation
  \item porMin. Minimum value of porosity (as decimal fraction, values of 0 to 1) distribution used in simulation
  \item porMax. Maximum value of porosity (as decimal fraction, values of 0 to 1) distribution used in simulation
  \item porMean. Mean value of porosity (as decimal fraction, values of 0 to 1) distribution used in simulation
  \item porSD. Standard deviation of  porosity (as decimal fraction, values of 0 to 1) distribution used in simulation
  \item fluDenMin. Minimum value of fluid density (g/cm3) distribution used in simulation
  \item fluDenMax. Maximum value of fluid density (g/cm3) distribution used in simulation
  \item fluDenMean. Mean value of fluid density (g/cm3) distribution used in simulation
  \item fluDenSD. Standard deviation of fluid density distribution (g/cm3) used in simulation
  \item uMin. Minimum value of uranium (ppm) distribution used in simulation
  \item uMax. Maximum value of uranium (ppm) distribution used in simulation
  \item uMean. Mean value of uranium (ppm) distribution used in simulation
  \item uSD. Standard deviation of uranium (ppm) distribution used in simulation
  \item thMin. Minimum value of thorium (ppm) distribution used in simulation
  \item thMax. Maximum value of thorium (ppm) distribution used in simulation
  \item thMean. Mean value of thorium (ppm) distribution used in simulation
  \item thSD. Standard deviation of thorium (ppm) distribution used in simulation
  \item kMin. Minimum value of potassium (WT%) distribution used in simulation
  \item kMax. Maximum value of potassium (WT%) distribution used in simulation
  \item kMean. Mean value of potassium (WT%) distribution used in simulation
  \item kSD. Standard deviation of potassium (WT%) distribution used in simulation
}
}
\references{
Radiolysis.
}
\keyword{datasets}
